requirements.txt
pandas
numpy
SQLAlchemy
psycopg2-binary
openpyxl
python-dotenv
matplotlib
